+++ Source/WebCore/platform/graphics/FontInlines.h
@@ -37,10 +37,6 @@
 
 namespace WebCore {
 
-#if ENABLE(OPENTYPE_VERTICAL)
-inline const OpenTypeVerticalData* Font::verticalData() const { return m_verticalData.get(); }
-#endif
-
 ALWAYS_INLINE FloatRect Font::boundsForGlyph(Glyph glyph) const
 {
     if (isZeroWidthSpaceGlyph(glyph))
