+++ Source/WebCore/platform/graphics/coretext/FontCoreText.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2005-2025 Apple Inc. All rights reserved.
+ * Copyright (C) 2005-2026 Apple Inc. All rights reserved.
  * Copyright (C) 2006 Alexey Proskuryakov
  *
  * Redistribution and use in source and binary forms, with or without
@@ -63,6 +63,11 @@ static inline bool caseInsensitiveCompare(CFStringRef a, CFStringRef b)
     return a && CFStringCompare(a, b, kCFCompareCaseInsensitive) == kCFCompareEqualTo;
 }
 
+static bool fontHasEitherTable(CTFontRef ctFont, unsigned tableTag1, unsigned tableTag2)
+{
+    return CTFontHasTable(ctFont, tableTag1) || CTFontHasTable(ctFont, tableTag2);
+}
+
 static bool fontHasVerticalGlyphs(CTFontRef font)
 {
     return fontHasEitherTable(font, kCTFontTableVhea, kCTFontTableVORG);
@@ -93,12 +98,7 @@ static bool isAhemFont(CFStringRef familyName)
     return familyName && caseInsensitiveCompare(familyName, CFSTR("Ahem"));
 }
 
-bool fontHasEitherTable(CTFontRef ctFont, unsigned tableTag1, unsigned tableTag2)
-{
-    return CTFontHasTable(ctFont, tableTag1) || CTFontHasTable(ctFont, tableTag2);
-}
-
-void Font::platformInit()
+void FontBase::platformInit()
 {
     auto constexpr syntheticBoldScaleFactor = 36.0f;
     m_syntheticBoldOffset = m_platformData.syntheticBold() ? (m_platformData.size() / syntheticBoldScaleFactor) : 0.f;
@@ -166,20 +166,6 @@ void Font::platformInit()
     m_shouldNotBeUsedForArabic = fontFamilyShouldNotBeUsedForArabic(familyName.get());
 #endif
 
-    CGFloat xHeight = 0;
-    if (m_platformData.size()) {
-        if (platformData().orientation() == FontOrientation::Horizontal) {
-            // Measure the actual character "x", since it's possible for it to extend below the baseline, and we need the
-            // reported x-height to only include the portion of the glyph that is above the baseline.
-            Glyph xGlyph = glyphForCharacter('x');
-            if (xGlyph)
-                xHeight = -CGRectGetMinY(platformBoundsForGlyph(xGlyph));
-            else
-                xHeight = CTFontGetXHeight(ctFont.get());
-        } else
-            xHeight = verticalRightOrientationFont().fontMetrics().xHeight().value_or(0);
-    }
-
     if (CTFontGetSymbolicTraits(ctFont.get()) & kCTFontTraitColorGlyphs) {
         if (RetainPtr cfBitVector = adoptCF(CTFontCopyColorGlyphCoverage(ctFont.get())))
             m_emojiType = SomeEmojiGlyphs { BitVector(cfBitVector.get()) };
@@ -193,7 +179,6 @@ void Font::platformInit()
     m_fontMetrics.setDescent(descent);
     m_fontMetrics.setCapHeight(capHeight);
     m_fontMetrics.setLineGap(lineGap);
-    m_fontMetrics.setXHeight(xHeight);
     m_fontMetrics.setLineSpacing(lineSpacing);
     m_fontMetrics.setUnderlinePosition(-CTFontGetUnderlinePosition(ctFont.get()));
     m_fontMetrics.setUnderlineThickness(CTFontGetUnderlineThickness(ctFont.get()));
@@ -225,6 +210,27 @@ void Font::platformCharWidthInit()
 
     // Fallback to a cross-platform estimate, which will populate these values if they are non-positive.
     initCharWidths();
+}
+
+void Font::platformCharHeightInit()
+{
+    CGFloat xHeight = 0;
+    if (m_platformData.size()) {
+        if (platformData().orientation() == FontOrientation::Horizontal) {
+            // Measure the actual character "x", since it's possible for it to extend below the baseline, and we need the
+            // reported x-height to only include the portion of the glyph that is above the baseline.
+            Glyph xGlyph = glyphForCharacter('x');
+            if (xGlyph)
+                xHeight = -CGRectGetMinY(platformBoundsForGlyph(xGlyph));
+            else {
+                RetainPtr ctFont = this->ctFont();
+                xHeight = CTFontGetXHeight(ctFont);
+            }
+        } else
+            xHeight = verticalRightOrientationFont().fontMetrics().xHeight().value_or(0);
+    }
+
+    m_fontMetrics.setXHeight(xHeight);
 }
 
 bool Font::variantCapsSupportedForSynthesis(FontVariantCaps fontVariantCaps) const
@@ -287,14 +293,45 @@ static void injectTrueTypeCoverage(int type, int selector, CTFontRef font, BitVe
     unionBitVectors(result, source.get());
 }
 
-bool Font::supportsOpenTypeAlternateHalfWidths() const
+static bool supportsOpenTypeFeature(CTFontRef font, CFStringRef featureTag)
+{
+    RetainPtr<CFArrayRef> features = adoptCF(CTFontCopyFeatures(font));
+    CFIndex featureCount = CFArrayGetCount(features.get());
+    for (CFIndex featureIndex = 0; featureIndex < featureCount; ++featureIndex) {
+        RetainPtr feature = static_cast<CFDictionaryRef>(CFArrayGetValueAtIndex(features.get(), featureIndex));
+        RetainPtr featureTypeIdentifier = static_cast<CFNumberRef>(CFDictionaryGetValue(feature.get(), kCTFontFeatureTypeIdentifierKey));
+        if (!featureTypeIdentifier)
+            continue;
+
+        int rawFeatureTypeIdentifier;
+        CFNumberGetValue(featureTypeIdentifier.get(), kCFNumberIntType, &rawFeatureTypeIdentifier);
+        if (rawFeatureTypeIdentifier != kTextSpacingType)
+            continue;
+
+        RetainPtr featureSelectors = static_cast<CFArrayRef>(CFDictionaryGetValue(feature.get(), kCTFontFeatureTypeSelectorsKey));
+        if (!featureSelectors)
+            continue;
+        auto selectorsCount = CFArrayGetCount(featureSelectors.get());
+        for (CFIndex selectorIndex = 0; selectorIndex < selectorsCount; ++selectorIndex) {
+            RetainPtr featureSelector = static_cast<CFDictionaryRef>(CFArrayGetValueAtIndex(featureSelectors.get(), selectorIndex));
+            RetainPtr openTypeTag = static_cast<CFStringRef>(CFDictionaryGetValue(featureSelector.get(), kCTFontOpenTypeFeatureTag));
+            if (!openTypeTag)
+                continue;
+            if (CFStringCompare(openTypeTag.get(), featureTag, 0) == kCFCompareEqualTo)
+                return true;
+        }
+    }
+    return false;
+}
+
+bool FontBase::supportsOpenTypeAlternateHalfWidths() const
 {
     if (m_supportsOpenTypeAlternateHalfWidths == SupportsFeature::Unknown)
         m_supportsOpenTypeAlternateHalfWidths = supportsOpenTypeFeature(protect(ctFont()).get(), CFSTR("halt")) ? SupportsFeature::Yes : SupportsFeature::No;
     return m_supportsOpenTypeAlternateHalfWidths == SupportsFeature::Yes;
 }
 
-bool Font::supportsSmallCaps() const
+bool FontBase::supportsSmallCaps() const
 {
     if (m_supportsSmallCaps == SupportsFeature::Unknown) {
         BitVector glyphsSupportedBySmallCaps;
@@ -306,7 +343,7 @@ bool Font::supportsSmallCaps() const
     return m_supportsSmallCaps == SupportsFeature::Yes;
 }
 
-bool Font::supportsAllSmallCaps() const
+bool FontBase::supportsAllSmallCaps() const
 {
     if (m_supportsAllSmallCaps == SupportsFeature::Unknown) {
         BitVector glyphsSupportedByAllSmallCaps;
@@ -320,7 +357,7 @@ bool Font::supportsAllSmallCaps() const
     return m_supportsAllSmallCaps == SupportsFeature::Yes;
 }
 
-bool Font::supportsPetiteCaps() const
+bool FontBase::supportsPetiteCaps() const
 {
     if (m_supportsPetiteCaps == SupportsFeature::Unknown) {
         BitVector glyphsSupportedByPetiteCaps;
@@ -332,7 +369,7 @@ bool Font::supportsPetiteCaps() const
     return m_supportsPetiteCaps == SupportsFeature::Yes;
 }
 
-bool Font::supportsAllPetiteCaps() const
+bool FontBase::supportsAllPetiteCaps() const
 {
     if (m_supportsAllPetiteCaps == SupportsFeature::Unknown) {
         BitVector glyphsSupportedByAllPetiteCaps;
@@ -513,37 +550,6 @@ RefPtr<Font> Font::platformCreateScaledFont(const FontDescription&, float scaleF
     RetainPtr<CTFontRef> scaledFont = adoptCF(CTFontCreateWithFontDescriptor(fontDescriptor.get(), size, nullptr));
 
     return createDerivativeFont(scaledFont.get(), size, m_platformData.orientation(), fontTraits, m_platformData.syntheticBold(), m_platformData.syntheticOblique(), m_platformData.widthVariant(), m_platformData.textRenderingMode(), protect(m_platformData.customPlatformData()).get(), m_platformData.metricsOverrides());
-}
-
-bool supportsOpenTypeFeature(CTFontRef font, CFStringRef featureTag)
-{
-    RetainPtr<CFArrayRef> features = adoptCF(CTFontCopyFeatures(font));
-    CFIndex featureCount = CFArrayGetCount(features.get());
-    for (CFIndex featureIndex = 0; featureIndex < featureCount; ++featureIndex) {
-        RetainPtr feature = static_cast<CFDictionaryRef>(CFArrayGetValueAtIndex(features.get(), featureIndex));
-        RetainPtr featureTypeIdentifier = static_cast<CFNumberRef>(CFDictionaryGetValue(feature.get(), kCTFontFeatureTypeIdentifierKey));
-        if (!featureTypeIdentifier)
-            continue;
-
-        int rawFeatureTypeIdentifier;
-        CFNumberGetValue(featureTypeIdentifier.get(), kCFNumberIntType, &rawFeatureTypeIdentifier);
-        if (rawFeatureTypeIdentifier != kTextSpacingType)
-            continue;
-
-        RetainPtr featureSelectors = static_cast<CFArrayRef>(CFDictionaryGetValue(feature.get(), kCTFontFeatureTypeSelectorsKey));
-        if (!featureSelectors)
-            continue;
-        auto selectorsCount = CFArrayGetCount(featureSelectors.get());
-        for (CFIndex selectorIndex = 0; selectorIndex < selectorsCount; ++selectorIndex) {
-            RetainPtr featureSelector = static_cast<CFDictionaryRef>(CFArrayGetValueAtIndex(featureSelectors.get(), selectorIndex));
-            RetainPtr openTypeTag = static_cast<CFStringRef>(CFDictionaryGetValue(featureSelector.get(), kCTFontOpenTypeFeatureTag));
-            if (!openTypeTag)
-                continue;
-            if (CFStringCompare(openTypeTag.get(), featureTag, 0) == kCFCompareEqualTo)
-                return true;
-        }
-    }
-    return false;
 }
 
 RefPtr<Font> Font::platformCreateHalfWidthFont() const
@@ -855,7 +861,7 @@ bool Font::isProbablyOnlyUsedToRenderIcons() const
     return !hasGlyphsForCharacterRange(platformFont.get(), ' ', '~', false) && !hasGlyphsForCharacterRange(platformFont.get(), 0x0600, 0x06FF, true);
 }
 
-const PAL::OTSVGTable& Font::otSVGTable() const
+const PAL::OTSVGTable& FontBase::otSVGTable() const
 {
     if (!m_otSVGTable) {
         if (auto tableData = adoptCF(CTFontCopyTable(protect(ctFont()).get(), kCTFontTableSVG, kCTFontTableOptionNoOptions)))
@@ -896,7 +902,7 @@ void Font::ComplexColorFormatGlyphs::set(Glyph glyph, bool value)
         m_bits.set(bitForValue(glyph));
 }
 
-bool Font::hasComplexColorFormatTables() const
+bool FontBase::hasComplexColorFormatTables() const
 {
     if (otSVGTable().table)
         return true;
@@ -909,7 +915,7 @@ bool Font::hasComplexColorFormatTables() const
     return false;
 }
 
-Font::ComplexColorFormatGlyphs& Font::glyphsWithComplexColorFormat() const
+Font::ComplexColorFormatGlyphs& FontBase::glyphsWithComplexColorFormat() const
 {
     if (!m_glyphsWithComplexColorFormat) {
         if (hasComplexColorFormatTables()) {
@@ -924,7 +930,7 @@ Font::ComplexColorFormatGlyphs& Font::glyphsWithComplexColorFormat() const
     return m_glyphsWithComplexColorFormat.value();
 }
 
-bool Font::glyphHasComplexColorFormat(Glyph glyphID) const
+bool FontBase::glyphHasComplexColorFormat(Glyph glyphID) const
 {
 #if HAVE(CORE_TEXT_GLYPHHASCOMPLEXCOLOR_FUNCTION)
     if (PAL::canLoad_CoreText_CTFontHasComplexColorFormatForGlyph())
@@ -945,7 +951,7 @@ bool Font::glyphHasComplexColorFormat(Glyph glyphID) const
     return false;
 }
 
-std::optional<BitVector> Font::findOTSVGGlyphs(std::span<const GlyphBufferGlyph> glyphs) const
+std::optional<BitVector> FontBase::findOTSVGGlyphs(std::span<const GlyphBufferGlyph> glyphs) const
 {
     auto table = otSVGTable().table;
     if (!table)
@@ -962,7 +968,7 @@ std::optional<BitVector> Font::findOTSVGGlyphs(std::span<const GlyphBufferGlyph>
     return result;
 }
 
-bool Font::hasAnyComplexColorFormatGlyphs(std::span<const GlyphBufferGlyph> glyphs) const
+bool FontBase::hasAnyComplexColorFormatGlyphs(std::span<const GlyphBufferGlyph> glyphs) const
 {
     auto& complexGlyphs = glyphsWithComplexColorFormat();
     if (!complexGlyphs.hasRelevantTables())
@@ -1049,7 +1055,7 @@ IPCFontData Font::toSerializableFont() const
 
 #if ENABLE(MULTI_REPRESENTATION_HEIC)
 
-MultiRepresentationHEICMetrics Font::metricsForMultiRepresentationHEIC() const
+MultiRepresentationHEICMetrics FontBase::metricsForMultiRepresentationHEIC() const
 {
     CGRect bounds = CTFontGetTypographicBoundsForAdaptiveImageProvider(ctFont(), nullptr);
